Allow float values for timeouts
(cherry picked from commit 6b37b45de1980b7fc64eb3206e41d86b8b886238)
(cherry picked from commit 3cf41879c0da7044f07b3a212f374f81494b18e1)
(cherry picked from commit ef490c828c868efd9c7c40e3c38867a704001a29)
(cherry picked from commit 5dc8341b8305510efef46bb4f030055b2f60a391)
(cherry picked from commit fde0617e5d989dd50b64442eee0feea66aef49e2)

+++ requests_retry_session/requests_retry_session.py
@@ -62,8 +62,8 @@ class RequestsRetryAdapterArgs(TypedDict, total=False):
     retries: int
     backoff_factor: float
     status_forcelist: Tuple[int, ...]
-    connect_timeout: int
-    read_timeout: int
+    connect_timeout: float
+    read_timeout: float
 
 
 def requests_session(adapter,
@@ -92,8 +92,8 @@ def requests_retry_adapter(
     retries: int
     backoff_factor: float
     status_forcelist: Tuple[int, ...]
-    connect_timeout: int
-    read_timeout: int
+    connect_timeout: float
+    read_timeout: float
     -> .timeout_http_adapter.TimeoutHTTPAdapter:
     """
     retry = RetryWithLogs(
